--- src/components/Filter/BiasList.tsx
@@ -1,28 +1,0 @@
-import React from "react";
-import { useQuery } from "react-query";
-import { useRecoilState } from "recoil";
-import { openedBiasAtom } from "../../state/atoms";
-import { fetchPeople } from "../../apis";
-import { StyledBiasList } from "../../styles/filterStyle";
-import Bias from "./Bias";
-
-function BiasList() {
-  const [openedBias] = useRecoilState(openedBiasAtom);
-
-  const { data: people } = useQuery(["people"], () => fetchPeople(), {
-    select: (data) => data?.filter((item) => openedBias.includes(item.id))
-  });
-
-  if (!people || !people.length) return null;
-  return (
-    <StyledBiasList>
-      {people?.map((person) =>
-        <Bias key={person.id}
-              id={person.id}
-              name={person.name}
-              profilePic={person.profilePic} />)}
-    </StyledBiasList>
-  );
-}
-
-export default BiasList;
