+++ src/pages/Main.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import EventList from "../components/EventList";
-import Filter from "../components/Filter";
 import Layout from "../components/layout";
+import Filter from "../components/main/Filter";
 import { StyledMain } from "../styles/mainStyle";
 
 function Main() {
